requirements.txt
#flask
#opencv-contrib-python
imutils
#opencv-python
requests
Pillow
picamera